refactor: extract tray icon and context menu into services/tray.ts
buildContextMenu() and createTrayIcon() extracted from main/index.ts
into main/services/tray.ts. Both receive AppState, StartupHandler,
paths, and a TrayCallbacks object (toggleMute, openEditorWindow,
onQuit) to decouple from the orchestrator's internal functions.

SYSTRAY_ICON constants also moved to tray.ts since they're only
used there.
